+++ components/frontend/src/subject/Subjects.js
@@ -1,32 +1,16 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { Segment } from 'semantic-ui-react';
 import { Subject } from './Subject';
 import { ReadOnlyOrEditable } from '../context/ReadOnly';
 import { CopyButton, AddButton, MoveButton } from '../widgets/Button';
 import { add_subject, copy_subject, move_subject } from '../api/subject';
 import { subject_options } from '../widgets/menu_options';
-
-function useDelayedRender() {
-  const [visible, setVisible] = useState(false);
-  useEffect(() => { setTimeout(setVisible, 50, true) }, []);
-  return visible;
-}
-
-function set_query_string(history, colums, hide_metrics) {
-  let queries = [];
-  if (colums.length > 0) {
-    queries.push(`hidden_columns=${colums.join(",")}`)
-  }
-  if (hide_metrics) {
-    queries.push(`hide_metrics_not_requiring_action=${hide_metrics}`)
-  }
-  history.replace({ search: queries.length > 0 ? "?" + queries.join("&") : "" })
-}
+import { useDelayedRender, useURLSearchQuery } from '../utils';
 
 export function Subjects(props) {
   const visible = useDelayedRender();
-  const [hideMetricsNotRequiringAction, setHideMetricsNotRequiringAction] = useState(props.hide_metrics_not_requiring_action);
-  const [hiddenColumns, setHiddenColumns] = useState(props.hidden_columns);
+  const [hideMetricsNotRequiringAction, setHideMetricsNotRequiringAction] = useURLSearchQuery(props.history, "hide_metrics_not_requiring_action", "boolean");
+  const [hiddenColumns, setHiddenColumns] = useURLSearchQuery(props.history, "hidden_columns", "array");
   const last_index = Object.keys(props.report.subjects).length - 1;
   return (
     <>
@@ -39,8 +23,8 @@ export function Subjects(props) {
             hideMetricsNotRequiringAction={hideMetricsNotRequiringAction}
             key={subject_uuid}
             last_subject={index === last_index}
-            setHiddenColumns={(columns) => { setHiddenColumns(columns); set_query_string(props.history, columns, hideMetricsNotRequiringAction) }}
-            setHideMetricsNotRequiringAction={(state) => { setHideMetricsNotRequiringAction(state); set_query_string(props.history, hiddenColumns, state) }}
+            setHiddenColumns={(columns) => setHiddenColumns(columns)}
+            setHideMetricsNotRequiringAction={(state) => setHideMetricsNotRequiringAction(state)}
             subject_uuid={subject_uuid}
           /> : null
       )}